File_chat_app/file_chatbot/requirements.txt
django==5.0.0
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==1.17.0
tiktoken==0.6.0
pypdf==4.0.1
python-docx==1.0.1
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3